LottoProfitCalculator 생성 : LottoChecker에서 수익률을 구하는 역할을 LottoProfitCalculator에 위임
