+++ src/test/java/lotto/domain/LottoCheckerTest.java
@@ -1,6 +1,5 @@
 package lotto.domain;
 
-import lotto.enums.LottoRank;
 import lotto.vo.LottoResult;
 import org.junit.Test;
 
@@ -44,40 +43,5 @@ public class LottoCheckerTest {
         assertThat(lottoResult.getSecond()).isEqualTo(1);
         assertThat(lottoResult.getThird()).isEqualTo(0);
         assertThat(lottoResult.getFourth()).isEqualTo(0);
-    }
-
-    @Test
-    public void 총_수익_계산() {
-        // given
-        long[] matchCounts = new long[Lotto.TOTAL_LOTTO_NUMBERS + 1];
-        matchCounts[LottoRank.FIRST.getMatchCount()]++;
-        matchCounts[LottoRank.THIRD.getMatchCount()]++;
-
-        LottoResult lottoResult = new LottoResult(matchCounts);
-
-        // when
-        long totalPrize = LottoChecker.getTotalPrizeMoney(lottoResult);
-
-        // then
-        assertThat(totalPrize)
-                .isEqualTo(LottoRank.FIRST.getPrizeMoney() + LottoRank.THIRD.getPrizeMoney());
-    }
-
-    @Test
-    public void 총_수익률_계산() {
-        // given
-        long[] matchCounts = new long[Lotto.TOTAL_LOTTO_NUMBERS + 1];
-        matchCounts[LottoRank.FIRST.getMatchCount()]++;
-        matchCounts[LottoRank.THIRD.getMatchCount()]++;
-
-        LottoResult lottoResult = new LottoResult(matchCounts);
-        long cost = 5_000;
-
-        // when
-        double totalProfitRate = LottoChecker.getTotalProfitRate(cost, lottoResult);
-
-        // then
-        assertThat(totalProfitRate)
-                .isEqualTo((LottoRank.FIRST.getPrizeMoney() + LottoRank.THIRD.getPrizeMoney()) / cost);
     }
 }
